--- BibliotecaDigital/src/CategoriaRecurso.java
@@ -1,5 +1,0 @@
-public enum CategoriaRecurso {
-    LIBRO,
-    REVISTA,
-    AUDIOLIBRO
-}
