--- BibliotecaDigital/src/EstadoRecurso.java
@@ -1,5 +1,0 @@
-public enum EstadoRecurso {
-    DISPONIBLE,
-    PRESTADO,
-    RESERVADO
-}
