+++ BibliotecaDigital/src/Main.java
@@ -1,3 +1,6 @@
+import consola.*;
+import notificaciones.*;
+
 public class Main {
     public static void main(String[] args) {
         ServicioNotificaciones notificador = new ServicioNotificacionesEmail(); //
