--- BibliotecaDigital/src/Prestable.java
@@ -1,5 +1,0 @@
-public interface Prestable {
-    void prestar() throws RecursoNoDisponibleException;
-    void devolver();
-    boolean estaPrestado();
-}
