--- BibliotecaDigital/src/RecursoDigital.java
@@ -1,9 +1,0 @@
-public interface RecursoDigital {
-    int getId();
-    String getTitulo();
-    CategoriaRecurso getCategoria();
-    EstadoRecurso getEstado();
-
-    void actualizarEstado(EstadoRecurso estado);
-    void mostrarInformacion();
-}
