--- BibliotecaDigital/src/RecursoNoDisponibleException.java
@@ -1,5 +1,0 @@
-public class RecursoNoDisponibleException extends Exception {
-    public RecursoNoDisponibleException(String mensaje) {
-        super(mensaje);
-    }
-}
