--- BibliotecaDigital/src/Renovable.java
@@ -1,3 +1,0 @@
-public interface Renovable {
-    void renovar();
-}
