--- BibliotecaDigital/src/ServicioNotificaciones.java
@@ -1,4 +1,0 @@
-// ServicioNotificaciones.java
-public interface ServicioNotificaciones {
-    void enviarNotificacion(String destino, String mensaje);
-}
