--- BibliotecaDigital/src/ServicioNotificacionesSMS.java
@@ -1,7 +1,0 @@
-// ServicioNotificacionesSMS.java
-public class ServicioNotificacionesSMS implements ServicioNotificaciones {
-    @Override
-    public void enviarNotificacion(String destino, String mensaje) {
-        System.out.println("[SMS a " + destino + "] " + mensaje);
-    }
-}
