--- BibliotecaDigital/src/UsuarioNoEncontradoException.java
@@ -1,5 +1,0 @@
-public class UsuarioNoEncontradoException extends Exception {
-    public UsuarioNoEncontradoException(String mensaje) {
-        super(mensaje);
-    }
-}
