+++ BibliotecaDigital/src/excepciones/RecursoNoDisponibleException.java
@@ -1,0 +1,7 @@
+package excepciones;
+
+public class RecursoNoDisponibleException extends Exception {
+    public RecursoNoDisponibleException(String mensaje) {
+        super(mensaje);
+    }
+}
