+++ BibliotecaDigital/src/excepciones/UsuarioNoEncontradoException.java
@@ -1,0 +1,7 @@
+package excepciones;
+
+public class UsuarioNoEncontradoException extends Exception {
+    public UsuarioNoEncontradoException(String mensaje) {
+        super(mensaje);
+    }
+}
