+++ BibliotecaDigital/src/modelo/CategoriaRecurso.java
@@ -1,0 +1,15 @@
+package modelo;
+
+import alertas.*;
+import excepciones.*;
+import gestores.*;
+import consola.*;
+import notificaciones.*;
+import reportes.*;
+import utils.*;
+
+public enum CategoriaRecurso {
+    LIBRO,
+    REVISTA,
+    AUDIOLIBRO
+}
