+++ BibliotecaDigital/src/modelo/EstadoRecurso.java
@@ -1,0 +1,15 @@
+package modelo;
+
+import alertas.*;
+import excepciones.*;
+import gestores.*;
+import consola.*;
+import notificaciones.*;
+import reportes.*;
+import utils.*;
+
+public enum EstadoRecurso {
+    DISPONIBLE,
+    PRESTADO,
+    RESERVADO
+}
