+++ BibliotecaDigital/src/notificaciones/ServicioNotificaciones.java
@@ -1,0 +1,5 @@
+package notificaciones;
+
+public interface ServicioNotificaciones {
+    void enviarNotificacion(String destino, String mensaje);
+}
