+++ BibliotecaDigital/src/notificaciones/ServicioNotificacionesEmail.java
@@ -1,0 +1,8 @@
+package notificaciones;
+
+public class ServicioNotificacionesEmail implements ServicioNotificaciones {
+    @Override
+    public void enviarNotificacion(String destino, String mensaje) {
+        System.out.println("[EMAIL a " + destino + "] " + mensaje);
+    }
+}
