+++ BibliotecaDigital/src/notificaciones/ServicioNotificacionesSMS.java
@@ -1,0 +1,8 @@
+package notificaciones;
+
+public class ServicioNotificacionesSMS implements ServicioNotificaciones {
+    @Override
+    public void enviarNotificacion(String destino, String mensaje) {
+        System.out.println("[SMS a " + destino + "] " + mensaje);
+    }
+}
