+++ BibliotecaDigital/src/utils/Renovable.java
@@ -1,0 +1,5 @@
+package utils;
+
+public interface Renovable {
+    void renovar();
+}
